Replacing CherryPy by Flask framework
WSGI used: Werkzeug.
