+++ centralreport/web/_server.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+"""
+    CentralReport - Server module
+        Manages internal web server
+
+    https://github.com/CentralReport
+"""
+
+#: @type app: Flask
+app = None
