+++ centralreport/web/static/css/style-responsive.css
