+++ centralreport/web/static/css/style.css
