+++ centralreport/web/static/img/check_18x18.png
