+++ centralreport/web/static/img/glyphicons-halflings-white.png
