+++ centralreport/web/static/img/glyphicons-halflings.png
