+++ centralreport/web/static/img/logos/Debian_logo.png
