+++ centralreport/web/static/js/cr.ajax.js
