+++ centralreport/web/templates/blocks/disks.block.tpl
