+++ centralreport/web/templates/json/full_check.json
