+++ tools/install_dev_tools.sh
@@ -18,7 +18,7 @@ source ../bash/functions.inc.sh
 #=============================================================================
 # List of libraries to install
 # Can be updated based on your needs
-declare -a LISTE='([0]="CherryPy" [1]="Jinja2")'
+declare -a LISTE='([0]="flask")'
 #=============================================================================
 clear
